feat: added proper test/exam type activity
- Implemented QuestionManagement component for managing exam questions, including adding, editing, deleting, importing, and exporting questions.
- Created WhitelistManagement component to manage student whitelists for exams, allowing selection and deselection of students.
- Integrated new ExamActivityModal for creating exams with relevant settings.
- Added form validation for exam creation using Zod and react-hook-form.

+++ apps/api/src/services/courses/activities/assignments.py
@@ -1681,7 +1681,7 @@ async def grade_assignment_submission(
         total += task_grade
 
     average = total / len(assignment_tasks)
-    rounded = int(round(average))
+    rounded = round(average)
 
     # Update the assignment user submission with the final rounded average
     assignment_user_submission.grade = rounded
